+++ CMakeLists.txt
@@ -84,7 +84,6 @@ add_library(GraphStreamingCC
   src/l0_sampling/sketch.cpp
   src/l0_sampling/update.cpp
   src/l0_sampling/pmshash.cpp
-  src/l0_sampling/fnv_1a.cpp
   src/util.cpp)
 add_dependencies(GraphStreamingCC GutterTree)
 target_link_libraries(GraphStreamingCC PUBLIC xxhash GutterTree)
@@ -100,7 +99,6 @@ add_library(GraphStreamingVerifyCC
   src/l0_sampling/sketch.cpp
   src/l0_sampling/update.cpp
   src/l0_sampling/pmshash.cpp
-  src/l0_sampling/fnv_1a.cpp
   src/util.cpp
   test/util/file_graph_verifier.cpp
   test/util/mat_graph_verifier.cpp)
@@ -141,7 +139,7 @@ if (BUILD_EXE)
 
   add_executable(xxhash_bench
     test/experiment/xxhash_speed.cpp
-    src/l0_sampling/fnv_1a.cpp)
+      )
   target_link_libraries(xxhash_bench PRIVATE xxhash)
   target_compile_definitions(xxhash_bench PRIVATE XXH_INLINE_ALL)
