+++ include/l0_sampling/pmshash.h
@@ -1,6 +1,6 @@
 #pragma once
 #include <cstdint>
 
-uint32_t mmp_hash_32(uint64_t x, uint64_t seed);
+uint64_t XXPMS64(uint64_t x, uint64_t seed);
 
-uint64_t mmp_hash_64(uint64_t x, uint64_t seed);
+uint32_t XXPMS32(uint64_t x, uint64_t seed);
