+++ test/experiment/mmp_hash.cpp
@@ -2,7 +2,7 @@
 #include <random>
 #include <vector>
 #include <bitset>
-#include "../../include/l0_sampling/xorshiro.h"
+#include "../../include/l0_sampling/pmshash.h"
 
 const uint64_t A1_1 = 10892479947228793040UL;
 const uint64_t A2_1 = 5324285833102856563UL;
@@ -147,14 +147,14 @@ void seed_expr(uint64_t (*hash_func)(uint64_t x, uint64_t seed)) {
 void plot_vals(int num_vals, uint64_t seed) {
   for (int i = 0; i < num_vals; ++i) {
     std::cout << std::dec << i << "\t";
-    auto val = xxxh3(i, seed);
+    auto val = XXPMS64(i, seed);
     std::cout << std::dec << (val >> 32) << "\t" << (uint32_t) val << "\n";  }
 }
 
 void plot_seeds(int num_vals, uint64_t x) {
   for (int i = 0; i < num_vals; ++i) {
     std::cout << std::dec << i << "\t";
-    auto val = xxxh3(x,i);
+    auto val = XXPMS64(x,i);
     std::cout << std::dec << (val >> 32) << "\t" << (uint32_t) val << "\n";
   }
 }
@@ -162,7 +162,7 @@ void plot_seeds(int num_vals, uint64_t x) {
 void collision_counter(int num_vals) {
   std::vector<uint32_t> vals(num_vals);
   for (int i = 0; i < num_vals; ++i) {
-    vals[i] = xxxh3(14, i) & 0xffffffff;
+    vals[i] = XXPMS64(14, i) & 0xffffffff;
   }
   int num_collisions = 0;
   for (int i = 0; i < num_vals; ++i) {
@@ -189,238 +189,3 @@ int main() {
 //  plot_vals(200, 602439);
   collision_counter(20000);
 }
-
-/*
---------------- COLLISION TEST RESULTS ---------------
-a1: 10892479947228793040, a2: 4032793241538373843, b: 15161517132367261381
-diff[1]: 17
-diff[2]: 17
-diff[3]: 16
-diff[4]: 17
-diff[5]: 15
-diff[6]: 16
-diff[7]: 17
-diff[8]: 17
-diff[9]: 16
-diff[10]: 15
-diff[11]: 16
-diff[12]: 16
-diff[13]: 15
-diff[14]: 17
-diff[15]: 15
---------------- COLLISION TEST RESULTS ---------------
-a1: 14826349932123903041, a2: 15419701087670201850, b: 11875562970292602379
-diff[1]: 17
-diff[2]: 17
-diff[3]: 15
-diff[4]: 17
-diff[5]: 15
-diff[6]: 16
-diff[7]: 16
-diff[8]: 17
-diff[9]: 19
-diff[10]: 15
-diff[11]: 15
-diff[12]: 15
-diff[13]: 15
-diff[14]: 16
-diff[15]: 15
---------------- TRAILING ZERO TEST RESULTS ---------------
-trail[0]: 49997230	 EXP 50000000
-trail[1]: 25003145	 EXP 25000000
-trail[2]: 12502016	 EXP 12500000
-trail[3]: 6248179	 EXP 6250000
-trail[4]: 3127404	 EXP 3125000
-trail[5]: 1561049	 EXP 1562500
-trail[6]: 781161	 EXP 781250
-trail[7]: 390469	 EXP 390625
-trail[8]: 194454	 EXP 195312
-trail[9]: 97294	 EXP 97656
-trail[10]: 49056	 EXP 48828
-trail[11]: 24438	 EXP 24414
-trail[12]: 12091	 EXP 12207
-trail[13]: 6142	 EXP 6103
-trail[14]: 2927	 EXP 3051
-trail[15]: 1479	 EXP 1525
-trail[16]: 751	 EXP 762
-trail[17]: 367	 EXP 381
-trail[18]: 168	 EXP 190
-trail[19]: 97	 EXP 95
-trail[20]: 36	 EXP 47
-trail[21]: 24	 EXP 23
-trail[22]: 9	 EXP 11
-trail[23]: 9	 EXP 5
-trail[24]: 2	 EXP 2
-trail[25]: 2	 EXP 1
-trail[26]: 1	 EXP 0
-trail[27]: 0	 EXP 0
-trail[28]: 0	 EXP 0
-trail[29]: 0	 EXP 0
-trail[30]: 0	 EXP 0
-trail[31]: 0	 EXP 100000000
-trail[32]: 0	 EXP 50000000
-trail[33]: 0	 EXP 25000000
-trail[34]: 0	 EXP 12500000
-trail[35]: 0	 EXP 6250000
-trail[36]: 0	 EXP 3125000
-trail[37]: 0	 EXP 1562500
-trail[38]: 0	 EXP 781250
-trail[39]: 0	 EXP 390625
-trail[40]: 0	 EXP 195312
-trail[41]: 0	 EXP 97656
-trail[42]: 0	 EXP 48828
-trail[43]: 0	 EXP 24414
-trail[44]: 0	 EXP 12207
-trail[45]: 0	 EXP 6103
-trail[46]: 0	 EXP 3051
-trail[47]: 0	 EXP 1525
-trail[48]: 0	 EXP 762
-trail[49]: 0	 EXP 381
-trail[50]: 0	 EXP 190
-trail[51]: 0	 EXP 95
-trail[52]: 0	 EXP 47
-trail[53]: 0	 EXP 23
-trail[54]: 0	 EXP 11
-trail[55]: 0	 EXP 5
-trail[56]: 0	 EXP 2
-trail[57]: 0	 EXP 1
-trail[58]: 0	 EXP 0
-trail[59]: 0	 EXP 0
-trail[60]: 0	 EXP 0
-trail[61]: 0	 EXP 0
-trail[62]: 0	 EXP 0
-trail[63]: 0	 EXP 100000000
-trail[64]: 0	 EXP 50000000
----------------- SEED: ZERO TEST RESULTS ----------------
-trail[0]: 5000709	 EXP 5000000
-trail[1]: 2498885	 EXP 2500000
-trail[2]: 1249793	 EXP 1250000
-trail[3]: 625539	 EXP 625000
-trail[4]: 312472	 EXP 312500
-trail[5]: 155914	 EXP 156250
-trail[6]: 78460	 EXP 78125
-trail[7]: 39053	 EXP 39062
-trail[8]: 19551	 EXP 19531
-trail[9]: 9877	 EXP 9765
-trail[10]: 4912	 EXP 4882
-trail[11]: 2449	 EXP 2441
-trail[12]: 1166	 EXP 1220
-trail[13]: 598	 EXP 610
-trail[14]: 300	 EXP 305
-trail[15]: 168	 EXP 152
-trail[16]: 84	 EXP 76
-trail[17]: 30	 EXP 38
-trail[18]: 23	 EXP 19
-trail[19]: 9	 EXP 9
-trail[20]: 2	 EXP 4
-trail[21]: 3	 EXP 2
-trail[22]: 3	 EXP 1
-trail[23]: 0	 EXP 0
-trail[24]: 0	 EXP 0
-trail[25]: 0	 EXP 0
-trail[26]: 0	 EXP 0
-trail[27]: 0	 EXP 0
-trail[28]: 0	 EXP 0
-trail[29]: 0	 EXP 0
-trail[30]: 0	 EXP 0
-trail[31]: 0	 EXP 10000000
-trail[32]: 0	 EXP 5000000
-trail[33]: 0	 EXP 2500000
-trail[34]: 0	 EXP 1250000
-trail[35]: 0	 EXP 625000
-trail[36]: 0	 EXP 312500
-trail[37]: 0	 EXP 156250
-trail[38]: 0	 EXP 78125
-trail[39]: 0	 EXP 39062
-trail[40]: 0	 EXP 19531
-trail[41]: 0	 EXP 9765
-trail[42]: 0	 EXP 4882
-trail[43]: 0	 EXP 2441
-trail[44]: 0	 EXP 1220
-trail[45]: 0	 EXP 610
-trail[46]: 0	 EXP 305
-trail[47]: 0	 EXP 152
-trail[48]: 0	 EXP 76
-trail[49]: 0	 EXP 38
-trail[50]: 0	 EXP 19
-trail[51]: 0	 EXP 9
-trail[52]: 0	 EXP 4
-trail[53]: 0	 EXP 2
-trail[54]: 0	 EXP 1
-trail[55]: 0	 EXP 0
-trail[56]: 0	 EXP 0
-trail[57]: 0	 EXP 0
-trail[58]: 0	 EXP 0
-trail[59]: 0	 EXP 0
-trail[60]: 0	 EXP 0
-trail[61]: 0	 EXP 0
-trail[62]: 0	 EXP 0
-trail[63]: 0	 EXP 10000000
-trail[64]: 0	 EXP 5000000
---------------- SEED: POPCNT TEST RESULTS ---------------
-pop[0]: 0
-pop[1]: 0
-pop[2]: 0
-pop[3]: 0
-pop[4]: 0
-pop[5]: 0
-pop[6]: 0
-pop[7]: 0
-pop[8]: 0
-pop[9]: 0
-pop[10]: 0
-pop[11]: 0
-pop[12]: 0
-pop[13]: 5
-pop[14]: 25
-pop[15]: 83
-pop[16]: 260
-pop[17]: 732
-pop[18]: 1937
-pop[19]: 4641
-pop[20]: 10654
-pop[21]: 22239
-pop[22]: 44063
-pop[23]: 79325
-pop[24]: 135652
-pop[25]: 218561
-pop[26]: 326320
-pop[27]: 458255
-pop[28]: 606778
-pop[29]: 753912
-pop[30]: 877864
-pop[31]: 962251
-pop[32]: 993098
-pop[33]: 962289
-pop[34]: 878973
-pop[35]: 754516
-pop[36]: 607139
-pop[37]: 457662
-pop[38]: 325353
-pop[39]: 217682
-pop[40]: 135711
-pop[41]: 79535
-pop[42]: 43622
-pop[43]: 22292
-pop[44]: 10640
-pop[45]: 4699
-pop[46]: 2044
-pop[47]: 790
-pop[48]: 260
-pop[49]: 101
-pop[50]: 26
-pop[51]: 8
-pop[52]: 3
-pop[53]: 0
-pop[54]: 0
-pop[55]: 0
-pop[56]: 0
-pop[57]: 0
-pop[58]: 0
-pop[59]: 0
-pop[60]: 0
-pop[61]: 0
-pop[62]: 0
-pop[63]: 0
-pop[64]: 0
- */
